pyproject.toml
[tool.poetry]
name = "ferris-elf"
version = "0.1.0"
description = "A bot to profile Rust code shared over Discord"
authors = ["proegssilb <proegssilb@gmail.com>"]
license = "Apache2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
docker = "^6.1.3"
"discord.py" = "^2.3.2"
SQLite4 = "^0.1.1"
dynaconf = "^3.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
